+++ client/src/App.tsx
@@ -1,7 +1,8 @@
-import { useEffect, useRef, useState } from 'react'
-import AddTodo from './components/AddTodo'
-import TodoItem from './components/TodoItem'
-import { Todo } from './interfaces/Todo.interface'
+import { useEffect, useRef, useState } from 'react';
+import AddTodo from './components/AddTodo';
+import TodoItem from './components/TodoItem';
+import { Todo } from './interfaces/Todo.interface';
+import { sortTasksByPriority, sortTasksByTime } from './components/api';
 
 const App = () => {
 
@@ -10,33 +11,6 @@ const App = () => {
   const [sortDataByPriority, setSortDataByPriority] = useState<boolean>(false);
   const [sortDataByTime, setSortDataByTime] = useState<boolean>(false);
   const isSorted = useRef(false);
-
-  const sortTasksByPriority = (data: Todo[]) => {
-    data = data.sort((t1, t2) => {
-      if (t1.priority > t2.priority) {
-          return 1;
-      } else if (t1.priority < t2.priority) {
-          return -1;
-      }
-      return 0;
-    });
-  }
-
-  const sortTasksByTime = (data: Todo[]) => {
-    const completedTasks = [];
-    const notCompletedTasks = [];
-    for (let i = 0; i < data.length; i++) {
-      if (data[i].completed) {
-        completedTasks.push(data[i]);
-      } else {
-        notCompletedTasks.push(data[i]);
-      }
-    }
-    const sortedCompletedTasks = completedTasks.sort((t1, t2) => {
-      return new Date(t1.completedTime).getTime() - new Date(t2.completedTime).getTime();
-    });
-    data = data.splice(0, data.length, ...sortedCompletedTasks.concat(notCompletedTasks));
-  }
 
   useEffect(() => {
     if (isSorted.current) {
@@ -60,7 +34,7 @@ const App = () => {
     <div>
       <h1>My Todo List</h1>
       <AddTodo dataUpdate={dataUpdate} setDataUpdate={setDataUpdate} />
-      {backendData ? <TodoItem backendData={backendData} dataUpdate={dataUpdate} setDataUpdate={setDataUpdate}
+      {backendData.length > 0 ? <TodoItem backendData={backendData} dataUpdate={dataUpdate} setDataUpdate={setDataUpdate}
                       sortDataByPriority={sortDataByPriority} setSortDataByPriority={setSortDataByPriority}
                       sortDataByTime={sortDataByTime} setSortDataByTime={setSortDataByTime} />
         : <h3>Nothing Yet...</h3>}
@@ -68,4 +42,4 @@ const App = () => {
   )
 }
 
-export default App
+export default App;
